feat : todo 관련 서버 액션 함수에 치명적이지 않은 에러는 이제 클라이언트에 출력됩니다.
- useActionState 훅을 통해 서버 액션에서 치밍적이지 않은 에러를 발견한
  경우 메세지를 클라이언트에 전달하여 사용자에게 에러 사항을 확인할 수
  있도록 합니다.

+++ src/actions/deleteTodo.ts
@@ -1,7 +1,8 @@
 "use server";
 
 import { connectDB } from "@/libs/database";
-import { ObjectId } from "mongodb";
+import { ITodo } from "@/types/schema";
+import { ObjectId, WithId } from "mongodb";
 import { revalidateTag } from "next/cache";
 
 export const deleteTodo = async (
@@ -12,6 +13,17 @@ export const deleteTodo = async (
   const todoid = formData.get("todoid") as string;
 
   const db = (await connectDB).db("next-todo-chart-cluster");
+  const todoDoc = await db
+    .collection<WithId<ITodo>>("todo")
+    .findOne({ _id: new ObjectId(todoid) });
+
+  if (!todoDoc) {
+    return { message: "조회 결과 해당 투두가 없습니다." };
+  }
+  if (todoDoc.userid !== userid) {
+    return { message: "투두를 작성한 사용자만 투두를 삭제할 수 있습니다." };
+  }
+
   await db.collection("todo").deleteOne({ _id: new ObjectId(todoid) });
   await db.collection("todos").updateOne(
     {
